refactor(open-mpm): split ctrl/ oversized files (closes #362)
Splits the 5 remaining oversized files in ctrl/ under the 500-line cap:
- supervisor.rs (1153) -> supervisor/{types,report,policy,runner,tests}.rs
- tests.rs (943) -> tests/{mod,config,state,command,pm_task,tools}_tests.rs
- pm_task.rs (830) -> pm_task/{helpers,dispatch/{history,persona}}.rs
- ctrl_turn.rs (548) -> ctrl_turn/{mod,dispatch}.rs
- repl.rs (528) -> repl/{mod,profile}.rs

Pure reorganization: public interface (CtrlSupervisor, SupervisorOutcome,
run_ctrl, run_ctrl_headless, run_pm_task_with_*) unchanged. All 1856 lib
tests pass; clippy + fmt clean.

Part of #356
